pyproject.toml
[project]
name = "mblm"
version = "0.0.1"
description = "Multiscale Byte Language Model"
authors = [
    { name = "Eric Egli", email = "eric.christian.egli@ibm.com" },
    { name = "Jannis Born", email = "jab@zurich.ibm.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "einops>=0.8.0",
    "filelock>=3.16.1",
    "mambapy>=1.2.0",
    "megabyte-pytorch>=0.3.5",
    "pillow>=11.0.0",
    "polars-lts-cpu[plot]>=1.8.2 ; sys_platform == 'linux'",
    "polars[plot]>=1.8.2 ; sys_platform != 'linux'",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "tabulate>=0.9.0",
    "tqdm>=4.66.5",
    "vegafusion[embed]>=1.6.9",
]

[tool.uv]
package = true
dev-dependencies = [
    "mypy>=1.11.2",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.6.8",
    "jupyter>=1.1.1",
    "types-pyyaml>=6.0.12.20240917",
    "nbqa>=1.9.0",
    "pre-commit>=3.8.0",
    "pytest-mock>=3.14.0",
    "types-tabulate>=0.9.0.20240106",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["tqdm.*", "mambapy.*", "mamba_ssm.*", "MEGABYTE_pytorch.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# import side effect for plotting with polars
module = ["hvplot.*"]
ignore_missing_imports = true

[tool.pytest]
mock_use_standalone_module = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--import-mode=importlib",
    "--cov-report=html:.coverage_html",
    "--cov-report=term-missing",
    "--cov=mblm",
]

[tool.ruff]
line-length = 100
target-version = "py310"
indent-width = 4
extend-include = ["*.ipynb"]

[tool.ruff.lint]
# see https://docs.astral.sh/ruff/rules/
# currently disabled rules can be reconsidered upon going public
select = [
    'ARG',
    # 'D100',
    # 'D101',
    # 'D102',
    # 'D103',
    # 'D104',
    # 'D105',
    'D106',
    # 'D107',
    'D300',
    'E4',
    'E7',
    'E721',
    'E9',
    'F',
    'I',
    'LOG001',
    'LOG002',
    'N8',
    'PTH',
    'PYI',
    'UP010',
    'UP011',
    'UP019',
    'UP032',
    'W',
]
# Due to license headers at top
ignore = ['E402', 'F404']

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
# use global line length
docstring-code-line-length = "dynamic"
